+++ IPDPS2021/collective-operations/kautz-conges.cpp
@@ -2,7 +2,7 @@
  * @ Author: Kien Pham
  * @ Create Time: 2021-10-05 11:33:06
  * @ Modified by: Kien Pham
- * @ Modified time: 2022-01-08 22:28:19
+ * @ Modified time: 2022-01-09 16:58:26
  * @ Description:
  */
 #include "kmpi.hpp"
@@ -373,7 +373,7 @@ int KMPI_Alltoallf(const float* sendbuf, int sendcount, MPI_Datatype sendtype,
     return (MPI_SUCCESS);
 }
 
-int KMPI_Allreducef(const float* sendbuf, float* recvbuf, int count,
+int KMPI_Allreducef(const float* sendbuf, float* recvbuf, size_t count,
     MPI_Datatype datatype, MPI_Op op, MPI_Comm comm){
     int rank;
     int size;
