+++ IPDPS2021/collective-operations/kmpi.hpp
@@ -2,7 +2,7 @@
  * @ Author: Kien Pham
  * @ Create Time: 2021-12-19 19:42:02
  * @ Modified by: Kien Pham
- * @ Modified time: 2022-01-08 22:26:26
+ * @ Modified time: 2022-01-09 16:59:25
  * @ Description:
  */
 
@@ -55,7 +55,7 @@ int KMPI_Alltoallf(const float* sendbuf, int sendcount, MPI_Datatype sendtype,
     float* recvbuf, int recvcount, MPI_Datatype recvtype,
     MPI_Comm comm);
 
-int KMPI_Allreducef(const float* sendbuf, float* recvbuf, int count,
+int KMPI_Allreducef(const float* sendbuf, float* recvbuf, size_t count,
     MPI_Datatype datatype, MPI_Op op, MPI_Comm comm);
 
 int KMPI_Init(int rank, int size);
